% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestions.R
\name{algo_similarity_based_reasoning}
\alias{algo_similarity_based_reasoning}
\title{Make suggestions using similarity based reasoning.}
\usage{
algo_similarity_based_reasoning(
  text_processed,
  suggestion_type,
  sim_name = "wordwise",
  min_aggregate_prob = 0.535,
  num_suggestions = 5,
  probabilities = occupationMeasurement::pretrained_models$similarity_based_reasoning,
  auxco = occupationMeasurement::auxco
)
}
\arguments{
\item{text_processed}{The processed user input.
Will be provided by get_job_suggestions.}

\item{suggestion_type}{Which type of suggestion to generate.
Will be provided by get_job_suggestions.}

\item{sim_name}{Which similarity measure to use.
Possible values are "wordwise" or "substring".}

\item{min_aggregate_prob}{The minimum aggregate probability for which to
return results. If the probability of suggestions is below this threshold
no suggestions will be returned, even if some have been found.}

\item{num_suggestions}{How many suggestions to return at the most.
Will be provided by get_job_suggestions.
Defaults to 5.}

\item{probabilities}{Trained probabilities to be used,
defaults to the one bundled with the package.}

\item{auxco}{Auxliary Classification (AuxCo) dataset to be used,
defaults to the one bundled with the package.}
}
\value{
A data.table with suggestions or NULL if no suggestions were found.
}
\description{
The Algorithm used here corresponds to Algorithm #10 in (Schierholz, 2019).
Note: This function should not be used directly, but rather as a step /
algorithm in get_job_suggestions.
}
\examples{
# Use with default settings
get_job_suggestions(
  "Arzt",
  steps = list(
    simbased_default = list(
      algorithm = algo_similarity_based_reasoning
    )
  )
)

# Use with substring similarity
get_job_suggestions(
  "Arzt",
  steps = list(
    simbased_substring = list(
      algorithm = algo_similarity_based_reasoning,
      parameters = list(
        sim_name = "substring"
      )
    )
  )
)
}
\references{
Schierholz, M. (2019). New Methods for Job and Occupation Classification (Ph.D. Thesis). University of Mannheim.
}
\seealso{
get_job_suggestions
}
