% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Start the occupation coding API.}
\usage{
api(
  start = TRUE,
  file = system.file("plumber", "api", "plumber.R", package = "occupationMeasurement")
)
}
\arguments{
\item{start}{Whether to immediately start the api. (Defaults to TRUE)}

\item{file}{Path to the \code{plumber.R} file describing the API.
Defaults to \code{plumber/api/plumber.R} within the installed package.
Refer to this file to understand how the API is implemented.}
}
\value{
A Plumber router
}
\description{
Start the occupation coding API.
}
\examples{
\dontrun{
api(start = TRUE)

router <- api(start = FALSE)
plumber::pr_run(router)
}
}
