% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Start the interactive shiny occupation coding app.}
\usage{
app(
  questionnaire = default_questionnaire(),
  app_settings = create_app_settings(),
  resource_dir = system.file("www", package = "occupationMeasurement"),
  ...
)
}
\arguments{
\item{questionnaire}{The questionnaire to load.
(Defaults to the questionnaire returned by \link{default_questionnaire}().)}

\item{app_settings}{The app_settings to use. Check the documentation for
create_app_settings to learn about the options.}

\item{resource_dir}{From which directory to static files e.g. styles.
If you want to load additional resources from outside the package, you
can do so with shiny::addResourcePath.}

\item{...}{Any additional parameters will be forwarded to shiny::shinyApp().}
}
\value{
A shiny app.
}
\description{
Start the interactive shiny occupation coding app.
}
\examples{
\dontrun{
app(
  app_settings = create_app_settings(
    save_to_file = TRUE
  )
)
}
}
\seealso{
default_questionnaire
}
