% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_helpers.R
\name{button_next}
\alias{button_next}
\alias{button_previous}
\title{Go to the next page}
\usage{
button_next(label = "Weiter")

button_previous(label = "Zurück")
}
\arguments{
\item{label}{What label the button should have.}
}
\value{
shiny Action Button
}
\description{
Buttons to navigate between pages.
}
\section{Functions}{
\itemize{
\item \code{button_previous()}: Go to the previous page

}}
\seealso{
\code{\link[=new_page]{new_page()}}
}
