% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{check_condition}
\alias{check_condition}
\title{Called internally by the shiny server.}
\usage{
check_condition(page, session, ...)
}
\arguments{
\item{session}{The shiny session}

\item{...}{All additional arguments are passed along}
}
\description{
Called internally by the shiny server.
}
\keyword{internal}
