% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestions.R
\name{convert_suggestions}
\alias{convert_suggestions}
\title{Convert suggestion from one suggestion format into the other}
\usage{
convert_suggestions(suggestions, from, to, suggestion_type_options = list())
}
\arguments{
\item{suggestions}{A data.table of suggestions as returned by one of the
prediction algorithms e.g. \code{\link[=algo_similarity_based_reasoning]{algo_similarity_based_reasoning()}}.}

\item{from}{The current suggestion_type in which suggestions are passed.}

\item{to}{The suggestion_type in which suggestions should be.}

\item{suggestion_type_options}{A list with options for generating
suggestions. Supported options:
- \code{datasets}: Pass specific datasets to be used whenn adding information
to predictions e.g. use a specific version of the kldb or auxco.
Supported datasets are: "auxco", "kldb". By default the datasets
bundled with this package are used.}
}
\value{
A data.table of suggestions in the "to" format
}
\description{
Convert suggestion from one suggestion format into the other
}
\keyword{internal}
