% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_settings.R
\name{create_app_settings}
\alias{create_app_settings}
\title{Create app_settings.}
\usage{
create_app_settings(
  suggestion_type = "auxco",
  require_id = FALSE,
  warn_before_leaving = FALSE,
  skip_followup_types = c(),
  save_to_file = TRUE,
  response_output_dir = file.path("output", "responses"),
  handle_data = NULL,
  get_job_suggestion_params = NULL,
  verbose = TRUE,
  .validate = TRUE
)
}
\arguments{
\item{suggestion_type}{Which type of suggestion to use / provide.
Possible options are "auxco" and "kldb-2010".}

\item{require_id}{Are user_ids required?
Defaults to FALSE}

\item{warn_before_leaving}{Should users be warned that their progress will
be lost upon leaving the site? Defaults to FALSE.}

\item{skip_followup_types}{A vector of strings corresponding to the
question_type of followup_question that should be skipped.}

\item{save_to_file}{Should responses be saved as files in
response_output_dir? Defaults to use the SAVE_TO_FILE environment variable
or TRUE if it is not set.}

\item{response_output_dir}{Path to the directory in which to store data
from the app. Defaults to \verb{./output/responses/}.}

\item{handle_data}{Callback function to handle data from the app.
This setting takes a function that get's passed 3 parameters:
table_name (A reference name indicating which data to save),
data (A dataframe of data to save), session (the user's current session).}

\item{get_job_suggestion_params}{List of parameters to pass to
get_job_suggestion. Refer to \code{\link[=get_job_suggestions]{get_job_suggestions()}} for a list of
supported parameters.}

\item{verbose}{Should additional output be printed when running?
Defaults to TRUE.}

\item{.validate}{Whether the created app_settings should be validated.
Defaults to TRUE.}
}
\value{
A list of app_settings.
}
\description{
This is the primary and most convenient way of configuring the app.
}
\examples{
app_settings <- create_app_settings(require_id = TRUE)
}
