% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{default_questionnaire}
\alias{default_questionnaire}
\title{The basic default questionnaire}
\usage{
default_questionnaire()
}
\value{
Default questionnaire.
}
\description{
View the function's code itself to see the used pages.
}
\examples{
library(occupationMeasurement)

# Inspect the code to create the default_questionnaire
print(default_questionnaire)

\dontrun{
# Run the app with the default_questionnaire
app(questionnaire = default_questionnaire())

# This is used by default within app
app()
}
}
