% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{demo_app}
\alias{demo_app}
\title{Run a demo version of the interactive app with additional explanations.}
\usage{
demo_app(...)
}
\arguments{
\item{...}{All paramaters will be passed on to \link{app}}
}
\value{
A shiny app.
}
\description{
This the same as \link{app}, using the \link{questionnaire_demo} instead
of the \link{questionnaire_web_survey}.
}
\examples{
\dontrun{
demo_app()
}
}
\seealso{
app, questionnaire_demo
}
