% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{demo_questionnaire}
\alias{demo_questionnaire}
\title{A demo questionnaire with additional explanations}
\usage{
demo_questionnaire()
}
\value{
Default questionnaire.
}
\description{
View the function's code itself to see the used pages.
Note, that this function has more complex code to show
the additional pages.
}
\examples{
library(occupationMeasurement)

# Inspect the code to create the demo_questionnaire
print(demo_questionnaire)

\dontrun{
# Run the app with the demo_questionnaire
app(questionnaire = demo_questionnaire())

# This is used by default within demo_app
demo_app()
}

}
