% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{execute_render}
\alias{execute_render}
\title{Called internally by the shiny server.}
\usage{
execute_render(page, session, run_before_output, ...)
}
\arguments{
\item{session}{The shiny session}

\item{run_before_output}{The output from \code{run_before}}

\item{...}{All additional arguments are passed along}
}
\description{
Called internally by the shiny server.
}
\keyword{internal}
