% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{execute_run_before}
\alias{execute_run_before}
\title{Called internally by the shiny server.}
\usage{
execute_run_before(page, session, input, output, ...)
}
\arguments{
\item{session}{The shiny session}

\item{...}{All additional arguments are passed along}
}
\description{
Called internally by the shiny server.
}
\keyword{internal}
