% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data}
\alias{get_data}
\title{Load a standard dataset, while supporting overriding by the user.}
\usage{
get_data(dataset_name, user_provided_data = list())
}
\arguments{
\item{dataset_name}{Name of the dataset. Currently supported values are
"auxco-1.2.x", "kldb-2010", "isco-08"}

\item{user_provided_data}{List of datasets provided by the user.}
}
\value{
The requested dataset.
}
\description{
Load a standard dataset, while supporting overriding by the user.
}
\keyword{internal}
