% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestions.R
\name{get_final_codes}
\alias{get_final_codes}
\title{Get the final occupation codes}
\usage{
get_final_codes(
  suggestion_id,
  followup_answers,
  code_type = c("isco_08", "kldb_10"),
  suggestion_type = "auxco"
)
}
\arguments{
\item{suggestion_id}{Id of the suggestion}

\item{followup_answers}{A named list of the question_ids with their
respective answers to the followup_questions.
Question ids correspond to list names, answers correspond to list values.}

\item{code_type}{Which type of codes should be returned.
Multiple codes can be returned at the same time.
Supported types of codes are "isco_08" and "kldb_10".
Defaults to "isco_08" and "kldb_10".}

\item{suggestion_type}{Which suggestion type is being used.
Only auxco-based suggestion_types are supported.}
}
\value{
A named list corresponding to the code_type(s) specified.
}
\description{
Get the final occupation codes
}
\examples{
get_final_codes(
  # First option for "Koch"
  "9079",
  followup_answers = list(
    # The first answer option in the followup questions
    "Q9079_1" = 1
  )
)
}
