% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestions.R
\name{get_followup_questions}
\alias{get_followup_questions}
\title{Get potential follow-up questions for a suggestion.}
\usage{
get_followup_questions(
  suggestion_id,
  tense = "present",
  suggestion_type = "auxco",
  auxco = occupationMeasurement::auxco
)
}
\arguments{
\item{suggestion_id}{Id of the suggestion}

\item{tense}{Which tense i.e. time to use for questions & answers,
this can be "present" or "past". Defaults to "present".}

\item{suggestion_type}{Which suggestion type is being used.
Only auxco-based suggestion_types are supported.}

\item{auxco}{Auxliary Classification (AuxCo) dataset to be used,
defaults to the one bundled with the package.}
}
\value{
List of followup questions and their answer options.
}
\description{
Get potential follow-up questions for a suggestion.
}
\examples{
# Get followup questions for "Post- und Zustelldienste"
get_followup_questions("1004")
}
