% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestions.R
\name{get_job_suggestions}
\alias{get_job_suggestions}
\title{Make coding suggestions based on a users free text input.}
\usage{
get_job_suggestions(
  text,
  suggestion_type = "auxco",
  num_suggestions = 5,
  suggestion_type_options = list(),
  score_thresholds = list(simbased_wordwise = 0.535, simbased_substring = 0.002),
  implausible_suggestion_threshold = 0,
  distinctions = TRUE,
  steps = list(simbased_wordwise = list(algorithm = algo_similarity_based_reasoning,
    parameters = list(sim_name = "wordwise")), simbased_substring = list(algorithm =
    algo_similarity_based_reasoning, parameters = list(sim_name = "substring")))
)
}
\arguments{
\item{text}{The raw text input from the user.}

\item{suggestion_type}{Which type of suggestion to use / provide.
Possible options are "auxco" and "kldb".}

\item{num_suggestions}{The maximum number of suggestions to show.
This is an upper bound and less suggestions may be returned.
Defaults to 5.}

\item{suggestion_type_options}{A list with options for generating
suggestions. Supported options:
- \code{datasets}: Pass specific datasets to be used whenn adding information
to predictions e.g. use a specific version of the kldb or auxco.
Supported datasets are: "auxco", "kldb". By default the datasets
bundled with this package are used.}

\item{score_thresholds}{A named list of thresholds between 0 and 1. Each
entry should correspond to one of the \code{steps}. Results from that step will
only be returned if the sum of their scores is equal to or greater than
the specified threshold. With a threshold of 0 results will always be
returned.}

\item{implausible_suggestion_threshold}{A threshold between 0 and 1 (usually
very small, default 0). Results from any step will only be returned if they
are greater than the specified threshold. Allows the removal of highly
implausible suggestions.}

\item{distinctions}{Whether or not to add additional distinctions to
similar occupational categories to the source code.
Defaults to TRUE.}

\item{steps}{A list with the algorithms to use and their parameters. Each
entry of the list should contain a nested list with two entries:
algorithm (the algorithm's function itself) and parameters (the parameters
to pass onto the algorithm). Each algorithm will also always have access
to a default set of three parameters:
\itemize{
\item text_processed: The input text after preprocessing
\item suggestion_type: Which type of suggestion to output
\item num_suggestions: How many suggestions shall be returned
These parameters must not be specified manually and will be provided
automatically instead.
Defaults to:
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  # try similarity "one word at most 1 letter different" first
  list(
    algorithm = algo_similarity_based_reasoning,
    parameters = list(
      sim_name = "wordwise",
      min_aggregate_prob = 0.535
    )
  ),
  # since everything else failed, try "substring" similarity
  list(
    algorithm = algo_similarity_based_reasoning,
    parameters = list(
      sim_name = "substring",
      min_aggregate_prob = 0.02
    )
  )
)
}\if{html}{\out{</div>}}}
}
\value{
A data.table with suggestions or NULL if no suggestions were found.
}
\description{
Make coding suggestions based on a users free text input.
}
\examples{
get_job_suggestions("Koch")

get_job_suggestions("Schlosser")
}
