% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{get_page_data}
\alias{get_page_data}
\title{Get questionnaire / page data.}
\usage{
get_page_data(session, page_id, key = NULL, default = NULL)
}
\arguments{
\item{session}{The shiny session}

\item{page_id}{The page for which to retrieve data.
Defaults to the page where data the function is being called from.}

\item{key}{The key for which to retrieve a value. (Optional)
If no key is provided, the page's whole data will be returned.}

\item{default}{A default value to return if the key or page is not
present in the questionnaire data.}
}
\value{
The page data value at the provided key or the whole page's data
if no key is provided.
}
\description{
Note that page data is \emph{not} automatically saved and you probably want
to use page$get_item_data instead.
}
\examples{
\dontrun{
# This code is expected to be run in e.g. run_before
occupationMeasurement:::get_page_data(session = session, key = "user_answer")
}
}
\seealso{
\code{\link[=get_item_data]{get_item_data()}}
}
\keyword{internal}
