% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{get_question_data}
\alias{get_question_data}
\title{Retrieve data for a question.}
\usage{
get_question_data(
  session,
  page_id,
  question_id = NULL,
  key = NULL,
  default = NULL
)
}
\arguments{
\item{session}{The shiny session}

\item{page_id}{The page for which to retrieve data.
Defaults to the page where data the function is being called from.}

\item{question_id}{The question for which to retrieve data.
This \emph{has} to be different for different questions on the same page.
Defaults to the page_id.}

\item{key}{The key for which to retrieve a value. (Optional)
If no key is provided, the page's whole data will be returned.}

\item{default}{A default value to return if the key or page is not
present in the questionnaire data.}
}
\value{
The question's data.
}
\description{
Each page in the questionnaire can have multiple questions on it.
}
\examples{
\dontrun{
# This code is expected to be run in e.g. run_before
page$get_question_data(
  session = session,
  key = "response_text",
  default = "alright"
)
}
}
