% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_saving.R
\name{get_responses}
\alias{get_responses}
\title{Convenience function to aggregate all saved results_overview files.}
\usage{
get_responses(app_settings = create_app_settings(save_to_file = TRUE))
}
\arguments{
\item{app_settings}{The app_settings configuration, should be the same as
used in \code{\link[=app]{app()}}.}
}
\value{
A combined data.table of user data (based on results_overview) or
NULL if there are no files.
}
\description{
Expects data to be saved as files.
}
\examples{
\dontshow{data.table::setDTthreads(1)}

\dontrun{
app_settings <- create_app_settings(save_to_file = TRUE)
if (interactive()) {
  get_responses(app_settings = app_settings)
}
}
}
