% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isco_08_en}
\alias{isco_08_en}
\title{Categories of the The International Standard Classification of Occupations - ISCO-08}
\format{
A data frame with 619 rows and 3 variables:
\describe{
\item{\code{code}}{character. Unique ISCO-08 identifier / code.}
\item{\code{label}}{character. Short label / title for the category.}
\item{\code{description}}{character. Detailed description of the category.}
}
}
\usage{
isco_08_en
}
\description{
Subset of the ISCO categories dataset bundled with the ESCO v1.1.
}
\details{
Source: https://esco.ec.europa.eu
More information on the ISCO-08: https://isco-ilo.netlify.app/en/isco-08/

This service uses the ESCO classification of the European Commission.
}
\keyword{datasets}
