% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{leaving_page_backwards}
\alias{leaving_page_backwards}
\title{Called internally by the shiny server when navigating to the previous page.}
\usage{
leaving_page_backwards(page, session, input, output, ...)
}
\arguments{
\item{session}{The shiny session}

\item{input}{The shiny input}

\item{...}{All additional arguments are passed along}
}
\description{
Called internally by the shiny server when navigating to the previous page.
}
\keyword{internal}
