% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_kldb}
\alias{load_kldb}
\title{Clean & Load KldB 2010 dataset.}
\usage{
load_kldb()
}
\value{
A cleaned / slimmed version of the KldB 2010.
}
\description{
Use load_kldb_raw() to load the whole dataset.
}
