% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_final}
\alias{page_final}
\title{A final page, showing instructions to close the window.}
\usage{
page_final(...)
}
\arguments{
\item{...}{All additional parameters are passed to \code{\link[=Page$new]{Page$new()}}}
}
\value{
A new \link{Page}
}
\description{
This page saves data in results_overview and marks the questionnaire as
complete.
}
