% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_first_freetext}
\alias{page_first_freetext}
\title{The first freetext question to show.}
\usage{
page_first_freetext(is_interview = FALSE, ...)
}
\arguments{
\item{is_interview}{Should the page show slightly different / additional
instructions and answer options for an interview that is conducted by
another person? Defaults to FALSE.}

\item{...}{All additional parameters are passed to \code{\link[=new_page]{new_page()}}}
}
\value{
A page object.
}
\description{
Here, the description of the job can be entered in an open freetext field.
}
\seealso{
\code{\link[=new_page]{new_page()}}
}
