% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_results}
\alias{page_results}
\title{Page showing the user's results}
\usage{
page_results(...)
}
\arguments{
\item{...}{All additional parameters are passed to \code{\link[=new_page]{new_page()}}}
}
\value{
A page object.
}
\description{
This page is only shown if the query parameter \code{show_results} is present.
}
\details{
This page saves data in results_overview and marks the questionnaire as
complete.
}
\seealso{
\code{\link[=new_page]{new_page()}}
}
