% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_second_freetext}
\alias{page_second_freetext}
\title{An optional, second free text question if the first didn't yield suggestions.}
\usage{
page_second_freetext(combine_input_with_first = TRUE, ...)
}
\arguments{
\item{combine_input_with_first}{Should input be combined with the
previous question?}

\item{...}{All additional parameters are passed to \code{\link[=Page$new]{Page$new()}}}
}
\value{
A new \link{Page}
}
\description{
If the first freetext question didn't provide satisfactory results,
ask for more details and try again.
}
