% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_select_suggestion}
\alias{page_select_suggestion}
\title{Display the generated suggestions for the user to pick one.}
\usage{
page_select_suggestion(...)
}
\arguments{
\item{...}{All additional parameters are passed to \code{\link[=Page$new]{Page$new()}}}
}
\value{
A new \link{Page}
}
\description{
Display the generated suggestions for the user to pick one.
}
