% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_welcome}
\alias{page_welcome}
\title{Welcome Page}
\usage{
page_welcome(
  title = "Herzlich Willkommen zum Modul zur automatischen Berufskodierung!",
  ...
)
}
\arguments{
\item{title}{The heading with which to greet participants.}

\item{...}{All additional parameters are passed to \code{\link[=Page$new]{Page$new()}}}
}
\value{
A new \link{Page}
}
\description{
Providing an introduction and greeting participants.
}
