% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pretrained_models}
\alias{pretrained_models}
\title{Pretrained ML models to be used with the package.}
\format{
A nested list with pretrained machine learning models:
\describe{
\item{\code{similarity_based_reasoning}}{list. Contains pretrained models to be used with \code{\link[=algo_similarity_based_reasoning]{algo_similarity_based_reasoning()}}.}
\item{\code{similarity_based_reasoning$wordwise}}{list. Contains the pretrained model to be used for providing suggestions using full wordwise matching.}
\item{\code{similarity_based_reasoning$substring}}{list. Contains the pretrained model to be used for providing suggestions using substring matching.}
}
}
\usage{
pretrained_models
}
\description{
Pretrained ML models to be used with the package.
}
\seealso{
algo_similarity_based_reasoning
}
\keyword{datasets}
