% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{questionnaire_demo}
\alias{questionnaire_demo}
\title{A demo questionnaire with additional explanations}
\usage{
questionnaire_demo()
}
\value{
A questionnaire for \code{\link[=app]{app()}} i.e. a list of pages.
}
\description{
View the function's code itself to see the used pages.
}
\details{
Note, that this function has more complex code to create
the additional pages.
}
\examples{
library(occupationMeasurement)

# Inspect the code to create the questionnaire_demo
print(questionnaire_demo)

\dontrun{
# Run the app with the questionnaire_demo
app(questionnaire = questionnaire_demo())
}

}
