% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{questionnaire_guided_interview}
\alias{questionnaire_guided_interview}
\title{A guided interview where an interviewer asks questions to an interviewee.}
\usage{
questionnaire_guided_interview()
}
\value{
A questionnaire for \code{\link[=app]{app()}} i.e. a list of pages.
}
\description{
View the function's code to see the used pages.
}
\examples{
library(occupationMeasurement)

# Inspect the code to create the questionnaire_guided_interview
print(questionnaire_guided_interview)

\dontrun{
# Run the app with the questionnaire_guided_interview
app(questionnaire = questionnaire_guided_interview())
}
}
