% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{questionnaire_interviewer_administered}
\alias{questionnaire_interviewer_administered}
\title{A guided interview where an interviewer asks questions to an interviewee.}
\usage{
questionnaire_interviewer_administered()
}
\value{
A questionnaire for \code{\link[=app]{app()}} i.e. a list of pages.
}
\description{
View the function's code to see the used pages.
}
\examples{
library(occupationMeasurement)

# Inspect the code to create the questionnaire_interviewer_administered
print(questionnaire_interviewer_administered)

\dontrun{
# Run the app with the questionnaire_interviewer_administered
app(questionnaire = questionnaire_interviewer_administered())
}
}
