% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{questionnaire_web_survey}
\alias{questionnaire_web_survey}
\title{A Web survey which participants can navigate themselves.}
\usage{
questionnaire_web_survey()
}
\value{
A questionnaire for \code{\link[=app]{app()}}, i.e. a list of pages.
}
\description{
The basic default questionnaire. View the function's code to see the
used pages. This function is meant as a template that can be changed to meet your requirements.
}
\examples{
library(occupationMeasurement)

# Inspect the code to create the questionnaire_web_survey
print(questionnaire_web_survey)

\dontrun{
# Run the app with the questionnaire_web_survey
app(questionnaire = questionnaire_web_survey())

# This is used by default within app
app()
}
}
