% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{set_page_data}
\alias{set_page_data}
\title{Set some values in the page/questionnaire data in the current session.}
\usage{
set_page_data(session, page_id, values)
}
\arguments{
\item{session}{The shiny session}

\item{values}{A named list of values to add / overwrite in the page data.
Values are added / overwritten based on the names provided in the list.}
}
\description{
Data is automatically linked to a page's page_id.
Note that page data is \emph{not} automatically saved and you probably want
to use page$set_question_data instead.
}
\examples{
\dontrun{
# This code is expected to be run in e.g. run_before
page$set_page_data(session = session, values = list(
  user_answer = "Some User Answer"
))
}
}
\seealso{
\code{\link[=set_question_data]{set_question_data()}}
}
\keyword{internal}
