% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{set_question_data}
\alias{set_question_data}
\title{Set / save data for a question.}
\usage{
set_question_data(
  session,
  page_id,
  question_id = NULL,
  question_text = NULL,
  response_text = NULL,
  response_id = NULL
)
}
\arguments{
\item{session}{The shiny session}

\item{page_id}{The page for which to retrieve data.}

\item{question_id}{The question for which to retrieve data.
This \emph{has} to be different for different questions on the same page.
Since most pages contain only a single question, \code{question_id} is set to "default" if missing.}

\item{question_text}{The question's text. (optional)}

\item{response_text}{The user's response in text form. (optional)}

\item{response_id}{The user's response as an id from a set of choices.
(optional)}
}
\description{
There can be multiple questions on any given page. The \code{question_text} is typically
saved in \code{run_before} and the reply (\code{response_text} and/or \code{response_id}) is
typically saved in \code{run_after}.
}
\examples{
\dontrun{
# This code is expected to be run in e.g. run_before
set_question_data(
  session = session,
  page_id = "example",
  question_text = "How are you?"
)

# This code is expected to be run in e.g. run_after
set_question_data(
  session = session,
  page_id = "example",
  response_id = 3,
  response_text = "I'm doing great!"
)
}
}
\seealso{
\code{\link[=get_question_data]{get_question_data()}}
}
