% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{validate_questionnaire}
\alias{validate_questionnaire}
\title{Validate (and sanitize the questionnaire)}
\usage{
validate_questionnaire(questionnaire, verbose)
}
\arguments{
\item{questionnaire}{The questionnaire passed to \code{\link[=app]{app()}}}

\item{verbose}{Should information about the questionnaire be printed?}
}
\value{
Sanitized questionnaire
}
\description{
Validate (and sanitize the questionnaire)
}
\keyword{internal}
