% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestions.R
\name{get_final_codes}
\alias{get_final_codes}
\title{Get the final occupation codes}
\usage{
get_final_codes(
  suggestion_id,
  followup_answers = list(),
  standardized_answer_levels = NULL,
  code_type = c("isco_08", "kldb_10"),
  suggestion_type = "auxco-1.2.x",
  suggestion_type_options = list()
)
}
\arguments{
\item{suggestion_id}{Id of the suggestion}

\item{followup_answers}{A named list of the question_ids with their
respective answers to the followup_questions.
Question ids correspond to list names, answers correspond to list values.}

\item{standardized_answer_levels}{A named list of standardized isco
answer levels. Names in the list correspond to the followup question type,
while values correspond to standard answer levels.
These can be used instead of some followup questions if
the information is available already from a different source.
Please note that standardized answer levels are \emph{not} available for all
question types. For a list of options please take a look at the
followup questions included in the auxco for example via
\code{occupationMeasurement::auxco$followup_questions}.}

\item{code_type}{Which type of codes should be returned.
Multiple codes can be returned at the same time.
Supported types of codes are "isco_08" and "kldb_10".
Defaults to "isco_08" and "kldb_10".}

\item{suggestion_type}{Which suggestion type is being used.
Only auxco-based suggestion_types are supported.}

\item{suggestion_type_options}{A list with options for generating
suggestions. Supported options:
- \code{datasets}: Pass specific datasets to be used whenn adding information
to predictions e.g. use a specific version of the kldb or auxco.
Supported datasets are: "auxco-1.2.x", "kldb-2010". By default the datasets
bundled with this package are used.}
}
\value{
A named list corresponding to the code_type(s) specified.
}
\description{
Get the final occupation codes
}
\examples{
get_final_codes(
  # Führungsaufgaben mit Personalverantwortung  bei der Lebensmittelherstellung
  "9076",
  followup_answers = list(
    # The first answer option in the first followup question
    "Q9076_1" = 1
  )
)

# The same, but using standardized answer levels
get_final_codes(
  # Führungsaufgaben mit Personalverantwortung  bei der Lebensmittelherstellung
  "9076",
  standardized_answer_levels = list(
    # A response corresponding to the standard ISCO Level "manager"
    "aufsicht" = "isco_manager"
  )
)
}
