% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestions.R
\name{get_suggestion_info}
\alias{get_suggestion_info}
\title{Get additional information for a suggestion id.}
\usage{
get_suggestion_info(suggestion_ids, suggestion_type = "auxco")
}
\arguments{
\item{suggestion_type}{Which suggestion type is being used.
Only auxco-based suggestion_types are supported.}

\item{suggestion_id}{Suggestion id(s) to retrieve information for.
This should typically be a character vector.}
}
\value{
Data table with information about the suggestion.
}
\description{
This includes a descriptive title, job descriptions etc.
To retrieve information regarding followup questions use
\link{get_followup_questions}.
}
\examples{
get_suggestion_info("9079")
}
\keyword{internal}
