% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isco_08_en}
\alias{isco_08_en}
\title{Categories of the The International Standard Classification of Occupations - ISCO-08}
\format{
A data frame with 619 rows and 3 variables:
\describe{
\item{\code{code}}{character. Unique ISCO-08 identifier / code.}
\item{\code{label}}{character. Short label / title for the category.}
\item{\code{description}}{character. Detailed description of the category.}
}
}
\usage{
isco_08_en
}
\description{
Categories from the International Standard Classification of Occupations - ISCO-08. ISCO-08 is a hierarchical classification, consisting of 10 (1-digit) major groups, 43 (2-digit) sub-major groups, 130 (3-digit) minor groups, and 436 (4-digit) unit groups, all of them included in this data set.
}
\details{
Source: https://esco.ec.europa.eu
This service uses the ESCO classification of the European Commission. The descriptions used here are taken from the ESCO classification (v1.1, Occupations pillar) of the European Commission, which is based on ISCO-08.

More information on the ISCO-08: https://isco-ilo.netlify.app/en/isco-08/, https://www.ilo.org/public/english/bureau/stat/isco/isco08/
}
\keyword{datasets}
