% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_feedback}
\alias{page_feedback}
\title{Page to receive feedback on how well the chosen suggestion fits}
\usage{
page_feedback(is_interview = FALSE, ...)
}
\arguments{
\item{is_interview}{Should the page show slightly different / additional
instructions and answer options for an interview that is conducted by
another person? Defaults to FALSE.}

\item{...}{All additional parameters are passed first passed on to
\code{\link[=page_choose_one_option]{page_choose_one_option()}} and then \code{\link[=new_page]{new_page()}}.}
}
\value{
A page object.
}
\description{
Page to receive feedback on how well the chosen suggestion fits
}
