% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_final}
\alias{page_final}
\title{A final page, showing instructions to close the window.}
\usage{
page_final(...)
}
\arguments{
\item{...}{All additional parameters are passed to \code{\link[=new_page]{new_page()}}}
}
\value{
A page object.
}
\description{
This page saves data in results_overview and marks the questionnaire as
complete.
}
\examples{
\dontshow{data.table::setDTthreads(1)}

\dontrun{
my_questionnaire <- list(
  page_final()
)
if (interactive()) {
  app(questionnaire = my_questionnaire)
}
}
}
\seealso{
\code{\link[=new_page]{new_page()}}
}
