% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_first_freetext}
\alias{page_first_freetext}
\title{The first freetext question to show.}
\usage{
page_first_freetext(...)
}
\arguments{
\item{...}{All additional parameters are passed to \code{\link[=Page$new]{Page$new()}}}
}
\value{
A new \link{Page}
}
\description{
Here, the description of the job can be entered in an open freetext field.
}
