% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_followup}
\alias{page_followup}
\title{Show potential followup questions to the user.}
\usage{
page_followup(index, ...)
}
\arguments{
\item{index}{The index of the followup question (1-based).
To show the first followup question (if there are any) use
page_followup(index = 1), to show a potential second followup question use
page_followup(index = 2).
For example \code{\link[=default_questionnaire]{default_questionnaire()}} uses
\verb{..., page_followup(index = 1), page_followup(index = 2), ...}}

\item{...}{All additional parameters are passed to \code{\link[=Page$new]{Page$new()}}}
}
\value{
A new \link{Page}
}
\description{
To disambiguate between similar occupations. Depending on the suggestion,
multiple followup questions can be shown.
}
