% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_results}
\alias{page_results}
\title{Page showing the user's results}
\usage{
page_results(...)
}
\arguments{
\item{...}{All additional parameters are passed to \code{\link[=new_page]{new_page()}}}
}
\value{
A page object.
}
\description{
This page is only meant for demonstration purposes. Users can see what they entered and which code was being saved. The page is only included in the \code{\link[=demo_questionnaire]{demo_questionnaire()}}, but not in the other questionnaire templates.
}
\seealso{
\code{\link[=new_page]{new_page()}}
}
