% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_pages.R
\name{page_welcome}
\alias{page_welcome}
\title{Welcome Page (optional)}
\usage{
page_welcome(
  title = "Herzlich Willkommen zum Modul zur automatischen Berufskodierung!",
  ...
)
}
\arguments{
\item{title}{The heading with which to greet participants.}

\item{...}{All additional parameters are passed to \code{\link[=new_page]{new_page()}}}
}
\value{
A page object.
}
\description{
Providing an introduction and greeting participants.
}
\examples{
my_questionnaire <- list(page_welcome)
if (interactive()){
  app(questionnaire = my_questionnaire)
}
}
\seealso{
\code{\link[=new_page]{new_page()}}
}
