% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaires.R
\name{questionnaire_interviewer_administered}
\alias{questionnaire_interviewer_administered}
\title{A questionnaire for interviewer-administered surveys}
\usage{
questionnaire_interviewer_administered(show_feedback_page = TRUE)
}
\arguments{
\item{show_feedback_page}{Show the \code{\link[=page_feedback]{page_feedback()}} to evaluate the fit of
the chosen suggestion.}
}
\value{
A questionnaire for \code{\link[=app]{app()}} i.e. a list of pages.
}
\description{
A questionnaire for Computer-assisted Interviewing (CAI), i.e. telephone interviewing or personal interviewing. In both modes, interviewer asks questions to an interviewee.
}
\details{
View the function's code to see the used pages. This function is meant as a template that can be changed to meet your requirements.
}
\examples{
library(occupationMeasurement)

# Inspect the code to create the questionnaire_interviewer_administered
print(questionnaire_interviewer_administered)

\dontrun{
# Run the app with the questionnaire_interviewer_administered
app(questionnaire = questionnaire_interviewer_administered())
}
}
